chore: Fix circular dependencies to avoid local import
